pyproject.toml
[project]
name = "video-auto-editor"
version = "0.1.0"
description = "基于 FastAPI 后端 + Streamlit 前端的视频自动剪辑系统"
readme = "SPEC.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.30.0",
    "ffmpeg-python>=0.2.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.27.0",
]


[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

